Refactoring, new methods for anomaly and commands.
