+++ flock_drone/mechanics/main.py
@@ -1,12 +1,14 @@
 """Handle main configuration for the drone."""
-import os
-import sys
+import json
 from hydra import Resource, SCHEMA
 from rdflib import Namespace
-import json
+
 from flock_drone.settings import CENTRAL_SERVER_NAMESPACE, DRONE_NAMESPACE
-from flock_drone.settings import DRONE_URL, CENTRAL_SERVER_URL
+from flock_drone.settings import CENTRAL_SERVER_URL
 from flock_drone.settings import IRI_CS, IRI_DRONE, DRONE_DEFAULT
+
+from flock_drone.mechanics.logs import send_http_api_log, gen_HttpApiLog
+
 
 global CENTRAL_SERVER, DRONE
 CENTRAL_SERVER = Namespace(CENTRAL_SERVER_NAMESPACE)
@@ -17,8 +19,6 @@ RES_CS = Resource.from_iri(IRI_CS)
 RES_DRONE = Resource.from_iri(IRI_DRONE)
 
 
-# Drone related methods
-# Status [Charging, Low Battery, Scanning, Off, Anomaly, Confirming]
 def get_drone_default():
     """Return the default drone object from settings."""
     return DRONE_DEFAULT
@@ -52,20 +52,9 @@ def get_controller_location():
         return "0,0"
 
 
-def get_drone_id():
-    """Return current drone id from drone server."""
-    drone = get_drone()
-    return int(drone["DroneID"])
-
-
-def get_drone_position():
-    """Return the drone position."""
-    drone = get_drone()
-    return drone["DroneState"]["Position"]
-
-
 def update_drone(drone):
     """Update the drone object on drone server."""
+    drone_identifier = drone["DroneID"]
     try:
         update_drone_ = RES_DRONE.find_suitable_operation(operation_type=SCHEMA.UpdateAction,
                                                           input_type=DRONE.Drone)
@@ -75,6 +64,9 @@ def update_drone(drone):
         return Resource.from_iri(resp['location'])
     except ConnectionRefusedError:
         raise ConnectionRefusedError("Connection Refused! Please check the drone server.")
+
+    http_api_log = gen_HttpApiLog("Drone %s" % (str(drone_identifier)), "POST DroneState", "Localhost")
+    send_http_api_log(http_api_log)
 
 
 def update_drone_at_controller(drone, drone_identifier):
@@ -92,137 +84,12 @@ def update_drone_at_controller(drone, drone_identifier):
     except:
         return {404: "Resource with Id %s not found!" % (id_,)}
 
-
-# Datastream related methods
-def gen_Datastream(temperature, position, drone_id):
-    """Generate a datastream objects."""
-    datastream = {
-        "@type": "Datastream",
-        "Temperature": temperature,
-        "Position": position,
-        "DroneID": drone_id,
-    }
-
-    return datastream
+    http_api_log = gen_HttpApiLog("Drone %s" % (str(drone_identifier)), "POST Drone", "Controller")
+    send_http_api_log(http_api_log)
 
 
-def gen_Anomaly(location):
-    """Generate an anomaly object."""
-    anomaly = {
-        "@type": "Anomaly",
-        "Location": location
-    }
-
-    return anomaly
-
-
-def update_datastream(datastream):
-    """Update the drone datastream on drone server."""
-    try:
-        update_datastream_ = RES_DRONE.find_suitable_operation(
-            operation_type=SCHEMA.UpdateAction, input_type=DRONE.Datastream)
-        resp, body = update_datastream_(datastream)
-        assert resp.status in [200, 201], "%s %s" % (resp.status, resp.reason)
-
-        return Resource.from_iri(resp['location'])
-    except ConnectionRefusedError:
-        raise ConnectionRefusedError("Connection Refused! Please check the drone server.")
-
-
-def get_datastream():
-    """Get the drone datastream from drone server."""
-    try:
-
-        get_datastream_ = RES_DRONE.find_suitable_operation(
-            operation_type=None, input_type=None, output_type=DRONE.Datastream)
-        resp, body = get_datastream_()
-        assert resp.status in [200, 201], "%s %s" % (resp.status, resp.reason)
-
-        datastream = json.loads(body.decode('utf-8'))
-        # remove extra contexts from datastream
-        datastream.pop("@context", None)
-        datastream.pop("@id", None)
-        return datastream
-    except ConnectionRefusedError:
-        raise ConnectionRefusedError("Connection Refused! Please check the drone server.")
-
-
-# Status related methods
-def gen_State(drone_id, battery, direction, position, sensor_status, speed):
-    """Generate a State objects."""
-    state = {
-        "@type": "State",
-        "DroneID": drone_id,
-        "Battery": battery,
-        "Direction": direction,
-        "Position": position,
-        "SensorStatus": sensor_status,
-        "Speed": speed,
-    }
-    return state
-
-
-def update_state(state):
-    """Update the drone state on drone server."""
-    drone = get_drone()
-    if int(drone["DroneID"]) == state["DroneID"]:
-        # Remove the DroneID key from state
-        state.pop("DroneID", None)
-
-        # Update the drone state
-        drone["DroneState"] = state
-        update_drone(drone)
-        print("Drone state updated successfully.")
-    else:
-        print("ERROR: DroneID %s not valid." % (state["DroneID"]))
-
-
-def get_state():
-    """Get the current drone state from the drone server."""
-    drone = get_drone()
-    drone_state = drone["DroneState"]
-    drone_state["DroneID"] = drone["DroneID"]
-
-    return drone_state
-
-# Command related methods
-
-
-def gen_Command(drone_id, state):
-    """Create a command entity."""
-    command = {
-        "@type": "Command",
-        "DroneID": drone_id,
-        "State": state
-    }
-    return command
-
-
-# Logs related Functions
-def gen_DroneLog(drone_id, log_string):
-    """Generate a Drone log object from log string."""
-    dronelog = {
-        "@type": "DroneLog",
-        "DroneID": drone_id,
-        "LogString": log_string
-    }
-    return dronelog
-
-
-def gen_HttpApiLog(source, action, target):
-    """Generate a Http Api Log object from action and target."""
-    httpapilog = {
-        "@type": "HttpApiLog",
-        "Subject": source,
-        "Predicate": action,
-        "Object": target
-    }
-    return httpapilog
-
-
-# Some general Functions
 def ordered(obj):
-    """Sort json dicts and lists within"""
+    """Sort json dicts and lists within."""
     if isinstance(obj, dict):
         return sorted((k, ordered(v)) for k, v in obj.items())
     if isinstance(obj, list):
@@ -232,7 +99,6 @@ def ordered(obj):
 
 
 if __name__ == "__main__":
-
     get_drone()
     print(update_drone(get_drone_default()))
     # print(get_drone_id())
