+++ flock_drone/mechanics/simulate.py
@@ -1,13 +1,16 @@
+"""Main control loop for drone."""
 import threading
 import random
-# from flock_drone.mechanics.manage_commands import get_command_collection
-from flock_drone.mechanics.main import get_drone, update_drone
+from flock_drone.mechanics.main import (get_drone, update_drone,
+                                        get_controller_location,
+                                        update_drone_at_controller)
+
+from flock_drone.mechanics.logs import (send_dronelog, send_http_api_log,
+                                        gen_DroneLog, gen_HttpApiLog)
+
+from flock_drone.mechanics.datastream import gen_Datastream, update_datastream, send_datastream
+from flock_drone.mechanics.anomaly import gen_Anomaly, send_anomaly
 from flock_drone.mechanics.distance import get_new_coordinates, gen_square_path, gen_drone_pos_limits
-from flock_drone.mechanics.main import update_drone_at_controller
-from flock_drone.mechanics.main import gen_Datastream, update_datastream, gen_Anomaly
-from flock_drone.mechanics.post_datastream import send_datastream
-from flock_drone.mechanics.main import gen_DroneLog, gen_HttpApiLog, get_controller_location
-from flock_drone.mechanics.post_logs import send_dronelog, send_http_api_log
 
 # Drone main Loop time settings
 global LOOP_TIME, ITERATOR
@@ -24,11 +27,28 @@ ITERATOR = 0
 def discharge_drone_battery(drone):
     """Handle drone battery discharging."""
     battery_level = drone["DroneState"]["Battery"]
+    drone_identifier = drone["DroneID"]
     if(int(battery_level) > 3):
         drone["DroneState"]["Battery"] = int(drone["DroneState"]["Battery"])-1
     else:
         # Battery level critical change drone status to OFF
         drone["DroneState"]["Status"] = "Off"
+
+    if int(battery_level) == 20:
+
+        dronelog = gen_DroneLog("Drone %s" % (str(drone_identifier),), "battery Low %s" % (str(drone["DroneState"]["Battery"])))
+        send_dronelog(dronelog)
+
+        http_api_log = gen_HttpApiLog("Drone %s" % (str(drone_identifier)), "PUT DroneLog", "Controller")
+        send_http_api_log(http_api_log)
+
+    elif int(battery_level) == 4:
+
+        dronelog = gen_DroneLog("Drone %s" % (str(drone_identifier)), "Battery level critical, will shutdown soon!")
+        send_dronelog(dronelog)
+
+        http_api_log = gen_HttpApiLog("Drone %s" % (str(drone_identifier)), "PUT DroneLog", "Controller")
+        send_http_api_log(http_api_log)
 
     return drone
 
@@ -113,95 +133,72 @@ def handle_drone_position(drone):
     """Handle the drone position changes."""
     drone_speed = float(drone["DroneState"]["Speed"])
     distance_travelled = calculate_dis_travelled(drone_speed, LOOP_TIME)
-
     drone_direction = str(drone["DroneState"]["Direction"])
+    drone_identifier = drone["DroneID"]
+    new_drone = update_drone_position(drone, distance_travelled, drone_direction)
 
-    drone = update_drone_position(drone, distance_travelled, drone_direction)
+    if(new_drone["DroneState"]["Direction"] != drone_direction):
 
-    return drone
+        dronelog = gen_DroneLog("Drone %s" % (str(drone_identifier),),
+                                "changed direction to %s" % (str(new_drone["DroneState"]["Direction"])))
+        send_dronelog(dronelog)
+
+        http_api_log = gen_HttpApiLog("Drone %s" % (str(drone_identifier)),
+                                      "PUT DroneLog", "Controller")
+        send_http_api_log(http_api_log)
+
+    return new_drone
 
 
-# Datastream related functions
-def gen_random_sensor_data():
+def gen_normal_sensor_data():
     """Generate random sensor data for drone datastream."""
-    drone_sensor_options = ["Normal", "Critical", "High"]
-    return random.choice(drone_sensor_options)
+    normal_range = range(25, 40)
+    return random.choice(normal_range)
 
 
 def gen_random_anomaly(drone):
-    """Generates an anomaly at random."""
-    option = random.choice([True, False])
-    if option:
-        anomaly = gen_Anomaly(drone["DroneState"]["Position"])
-
+    """Generate an anomaly at random."""
+    global ITERATOR
+    ITERATOR += 1
+    # 1/3 chance of anomaly every ten iterations
+    if ITERATOR % 10 == 0:
+        ITERATOR = 0
+        gen_random_anomaly()
+        option = random.choice([False, True, False])
+        if option:
+            anomaly = gen_Anomaly(drone["DroneState"]["Position"])
+            return anomaly
+    return None
 
 
 def main():
     """15 second time loop for drone."""
-    global ITERATOR
-    ITERATOR += 1
-    if ITERATOR % 10 == 0:
-        gen_random_anomaly()
-
     drone = get_drone()
 
-    ## handle the drone battery change
-    drone_identifier = drone["DroneID"]
+    # Handle positions and battery change
     drone = handle_drone_battery(drone)
-    drone_battery = drone["DroneState"]["Battery"]
-    drone_direction = drone["DroneState"]["Direction"]
-
-    ## Handle drone position change
     drone = handle_drone_position(drone)
 
-    if(drone["DroneState"]["Direction"] != drone_direction):
-        ## Generate and send DroneLog
-        dronelog = gen_DroneLog("Drone %s" %(str(drone_identifier),), "changed direction to %s" %(str(drone["DroneState"]["Direction"])))
-        send_dronelog(dronelog)
-        ## Generate and send HttpApiLog
-        http_api_log = gen_HttpApiLog("Drone %s" %(str(drone_identifier)),"PUT DroneLog", "Controller")
-        send_http_api_log(http_api_log)
+    drone_identifier = drone["DroneID"]
 
-    ## Update the drone at central contoller.
-    ## Send Drone battery and Http Api logs
-    if int(drone_battery == 20):
-        ## Generate and send DroneLog
-        dronelog = gen_DroneLog("Drone %s" %(str(drone_identifier),), "battery Low %s" %(str(drone["DroneState"]["Battery"])))
-        send_dronelog(dronelog)
-        ## Generate and send HttpApiLog
-        http_api_log = gen_HttpApiLog("Drone %s" %(str(drone_identifier)),"PUT DroneLog", "Controller")
-        send_http_api_log(http_api_log)
-
-    elif int(drone_battery == 4):
-        ## Generate and send DroneLog
-        dronelog = gen_DroneLog("Drone %s" %(str(drone_identifier)), "Battery level critical, will shutdown soon!")
-        send_dronelog(dronelog)
-        ## Generate and send HttpApiLog
-        http_api_log = gen_HttpApiLog("Drone %s" %(str(drone_identifier)),"PUT DroneLog", "Controller")
-        send_http_api_log(http_api_log)
+    anomaly = gen_random_anomaly()
+    if anomaly is not None:
+        send_anomaly(anomaly)
 
     if int(drone_identifier) != -1000:
         update_drone_at_controller(drone, drone_identifier)
-        ## Generate and send HttpApiLog
-        http_api_log = gen_HttpApiLog("Drone %s" %(str(drone_identifier)),"POST Drone", "Controller")
-        send_http_api_log(http_api_log)
-    ## update the drone locally
-    update_drone(drone)
-    ## Generate and send HttpApiLog
-    http_api_log = gen_HttpApiLog("Drone %s" %(str(drone_identifier)),"POST DroneState", "Localhost")
-    send_http_api_log(http_api_log)
 
+    # update the drone locally
+    update_drone(drone)
+    # Generate and send HttpApiLog
 
     # Handle sensor datastream
-    datastream = gen_Datastream(gen_random_sensor_data(), drone["DroneState"]["Position"], drone_identifier)
     if int(drone_identifier != -1000):
+        datastream = gen_Datastream(gen_normal_sensor_data(), drone["DroneState"]["Position"], drone_identifier)
         send_datastream(datastream)
-        ## Generate and send HttpApiLog
-        http_api_log = gen_HttpApiLog("Drone %s" %(str(drone_identifier)),"PUT Datastream", "Controller")
-        send_http_api_log(http_api_log)
-    update_datastream(datastream)
+        update_datastream(datastream)
 
-    # call main() again in 60 LOOP_TIME
+    # call main() again in LOOP_TIME
     threading.Timer(LOOP_TIME, main).start()
